HMW-45 Renamed the Map component to HmwMap, so that it is more clear the difference between the internal component and the Esri one.
