+++ app/client/src/components/pages/Actions/ActionsMap.js
@@ -9,7 +9,7 @@ import SimpleFillSymbol from '@arcgis/core/symbols/SimpleFillSymbol';
 import SimpleMarkerSymbol from '@arcgis/core/symbols/SimpleMarkerSymbol';
 import Viewpoint from '@arcgis/core/Viewpoint';
 // components
-import Map from 'components/shared/Map';
+import HmwMap from 'components/shared/HmwMap';
 import MapLoadingSpinner from 'components/shared/MapLoadingSpinner';
 import MapErrorBoundary from 'components/shared/ErrorBoundary/MapErrorBoundary';
 // styled components
@@ -347,7 +347,7 @@ function ActionsMap({ layout, unitIds, onLoad }: Props) {
 
   return (
     <Container data-testid="hmw-actions-map">
-      <Map layers={layers} />
+      <HmwMap layers={layers} />
       {mapView && mapLoading && <MapLoadingSpinner />}
     </Container>
   );
